Agents and Workflows/deep_memory_researcher_mem0/extra/deep_memory_researcher/requirements.txt
# Deep Memory Researcher Dependencies

# Core dependencies from existing agents
mem0ai
python-dotenv
rich
camel-ai
pydantic

# Additional utilities 
python-dateutil